gh-24: Implements basic parsing of compositions and Environment creation.
