+++ flocking_cpal/examples/print_audio_tree/main.rs
@@ -1,5 +1,5 @@
 extern crate flocking_cpal;
 
 fn main() {
-    flocking_cpal::print_audio_system_tree();
+    flocking_cpal::utils::print_audio_tree();
 }
