+++ flocking_cpal/examples/run_composition/test_composition.json
@@ -1,0 +1,5 @@
+{
+    "environment": {
+        "sample_rate": 44100
+    }
+}
