+++ flocking_cpal/src/env.rs
@@ -1,0 +1,10 @@
+// TODO: Export Environment to the root of flocking_cpal.
+
+#[derive(Debug)]
+pub struct Environment {}
+
+impl Environment {
+    pub fn new(settings: flocking::EnvironmentSettings) -> Environment {
+        Environment {}
+    }
+}
